requirements.txt
# LiveKit Core and Agents Framework
livekit>=1.0.23
livekit-agents>=1.3.10
livekit-api>=1.1.0
livekit-protocol>=1.1.1

# LiveKit Plugins
livekit-plugins-google>=1.3.10
livekit-plugins-silero>=1.3.10

# Environment and Configuration
python-dotenv>=1.0.0

# Additional utilities (already installed with agents)
# aiohttp, asyncio, etc. are included as dependencies